Adapters are now proxied to apply filters to methods inputs/outputs

+++ test/adapters/mongo.js
@@ -3,16 +3,19 @@
 const AdapterTestUtils = require('./utils');
 
 const ADAPTER_LABEL = 'mongo';
-const dataSource = AdapterTestUtils.createDataSource(ADAPTER_LABEL, {
-	database: 'diaspora_test',
-	appname: 'DiasporaTest',
-	username: 'admin',
-});
+const adapterConfig = getConfig(ADAPTER_LABEL);
+if(!adapterConfig.database){
+	return it('MongoDB adapter unconfigured', function(){
+		this.skip();
+	});
+}
+
+const dataSource = AdapterTestUtils.createDataSource(ADAPTER_LABEL, l.assign({ appname: 'DiasporaTest' }, adapterConfig));
 it('Wait for purge of mongodb collection', () => {
-	dataSource.waitReady().then(() => {
+	return dataSource.waitReady().then(() => {
 		const collection = dataSource.db.collection('test');
 		return collection.drop();
-	});
+	}).catch(e => console.error(e));
 });
 AdapterTestUtils.checkSpawnedAdapter(ADAPTER_LABEL, 'Mongo');
 AdapterTestUtils.checkEachStandardMethods(ADAPTER_LABEL);
